rename nav-structure
changed _ to -
